rename package to username
