+++ app/build.gradle.kts
@@ -5,15 +5,15 @@ plugins {
 }
 
 android {
-    namespace = "com.example.malankaraorthodoxliturgica"
+    namespace = "com.paradox543.malankaraorthodoxliturgica"
     compileSdk = 35
 
     defaultConfig {
-        applicationId = "com.example.malankaraorthodoxliturgica"
+        applicationId = "com.paradox543.malankaraorthodoxliturgica"
         minSdk = 24
         targetSdk = 35
         versionCode = 1
-        versionName = "1.0"
+        versionName = "0.1.0"
 
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
     }
