+++ app/src/androidTest/java/com/paradox543/malankaraorthodoxliturgica/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.malankaraorthodoxliturgica
+package com.paradox543.malankaraorthodoxliturgica
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.example.malankaraorthodoxliturgica", appContext.packageName)
+        assertEquals("com.paradox543.malankaraorthodoxliturgica", appContext.packageName)
     }
 }
