+++ app/src/main/java/com/paradox543/malankaraorthodoxliturgica/MainActivity.kt
@@ -1,28 +1,20 @@
-package com.example.malankaraorthodoxliturgica
+package com.paradox543.malankaraorthodoxliturgica
 
-//import com.example.malankaraorthodoxliturgica.view.setAppLocale
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.activity.viewModels
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBar
-import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import com.example.malankaraorthodoxliturgica.model.DataStoreManager
-import com.example.malankaraorthodoxliturgica.model.PrayerRepository
-import com.example.malankaraorthodoxliturgica.ui.theme.MalankaraOrthodoxLiturgicaTheme
-import com.example.malankaraorthodoxliturgica.view.navigation.NavGraph
-import com.example.malankaraorthodoxliturgica.viewmodel.PrayerViewModel
-import com.example.malankaraorthodoxliturgica.viewmodel.PrayerViewModelFactory
+import com.paradox543.malankaraorthodoxliturgica.model.DataStoreManager
+import com.paradox543.malankaraorthodoxliturgica.model.PrayerRepository
+import com.paradox543.malankaraorthodoxliturgica.ui.theme.MalankaraOrthodoxLiturgicaTheme
+import com.paradox543.malankaraorthodoxliturgica.view.navigation.NavGraph
+import com.paradox543.malankaraorthodoxliturgica.viewmodel.PrayerViewModel
+import com.paradox543.malankaraorthodoxliturgica.viewmodel.PrayerViewModelFactory
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
