+++ app/src/main/java/com/paradox543/malankaraorthodoxliturgica/view/QurbanaScreen.kt
@@ -1,7 +1,6 @@
-package com.example.malankaraorthodoxliturgica.view
+package com.paradox543.malankaraorthodoxliturgica.view
 
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
@@ -9,7 +8,6 @@ import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
@@ -17,31 +15,29 @@ import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
-import com.example.malankaraorthodoxliturgica.viewmodel.PrayerViewModel
+import com.paradox543.malankaraorthodoxliturgica.viewmodel.PrayerViewModel
 
 @Composable
-fun HomeScreen(navController: NavController, prayerViewModel: PrayerViewModel) {
-    val translations = prayerViewModel.loadTranslations()
-    val categories = prayerViewModel.getCategories()
+fun QurbanaScreen(navController: NavController, prayerViewModel: PrayerViewModel){
+    val sections = prayerViewModel.getQurbanaSections()
     val selectedFontSize by prayerViewModel.selectedFontSize.collectAsState()
-    prayerViewModel.setTopBarKeys(listOf("malankara"))
-    LazyColumn(modifier = Modifier.fillMaxSize()) {
-        items(categories) { category ->
+
+    LazyColumn(modifier = Modifier) {
+        items(sections) { section ->
             Card(
                 modifier = Modifier
                     .fillMaxWidth()
-                    .padding(8.dp)
-                    .clickable { navController.navigate("prayer_list/$category") },
+                    .clickable { navController.navigate("dummy") }
+                    .padding(8.dp),
                 shape = RoundedCornerShape(8.dp),
                 elevation = CardDefaults.cardElevation(4.dp)
             ) {
                 Text(
-                    text = translations[category] ?: "",
+                    text = section,
                     fontSize = selectedFontSize,
-                    modifier = Modifier.padding(16.dp),
-                    style = MaterialTheme.typography.titleMedium
-                )
+                    modifier = Modifier.padding(16.dp))
             }
         }
     }
+
 }
