+++ app/src/test/java/com/paradox543/malankaraorthodoxliturgica/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.example.malankaraorthodoxliturgica
+package com.paradox543.malankaraorthodoxliturgica
 
 import org.junit.Test
